pyproject.toml
[project]
name = "wordparse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "openai>=2.15.0",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
]
